Actualicé estilo lista de botones
