+++ src/sass/_fondos.scss
@@ -1,0 +1,48 @@
+.bg-orange {
+  background-color: $bg-orange;
+  &:hover{
+    background-color: $bg-orange-hover;
+  }
+}
+
+.bg-green {
+  background-color: $bg-green;
+  &:hover{
+    background-color: $bg-green-hover;
+  }
+}
+
+.bg-pink {
+  background-color: $bg-pink;
+  &:hover{
+    background-color: $bg-pink-hover;
+  }
+}
+
+.bg-blue {
+  background-color: $bg-blue;
+  &:hover{
+    background-color: $bg-blue-hover;
+  }
+}
+
+.bg-purple {
+  background-color: $bg-purple;
+  &:hover{
+    background-color: $bg-purple-hover;
+  }
+}
+
+.bg-red {
+  background-color: $bg-red;
+  &:hover{
+    background-color: $bg-red-hover;
+  }
+}
+
+.bg-teal {
+  background-color: $bg-teal;
+  &:hover{
+    background-color: $bg-teal-hover;
+  }
+}
