+++ src/sass/index.scss
@@ -2,6 +2,7 @@
 @import './plan-de-estudio.scss';
 @import './barra-progreso.scss';
 @import './shared.scss';
+@import './fondos.scss';
 
 * {
   color: white;
