+++ src/sass/plan-de-estudio.scss
@@ -1,87 +1,15 @@
 .titulo-tabla {
   display: flex;
-  /* float: none; */
-  /* float: right; */
   float: right;
-  /* justify-content: center; */
-  /* align-self: center  */
-  /* align-self: center; */
-}
-
-.boton-color {
-  border: 1px solid;
-  color: white;
-  height: 50px; 
-  width: 50px;
 }
 
 .semester {
-  /* display: flex; */
-  /* width: 100/9%; */
   color: white;
-  /* padding: 10px; */
-  /* margin: 5px; */
-  /*background-color: white;*/
   text-align: center;
-  /*border: white; */
-  // padding: 1rem;
-  // margin: 5px;
-  
-  /* text-align: center;
-  border: ridge white; */
 }
 
 .boton-seleccionado {
   border: 3px solid;
-}
-
-.bg-orange {
-  background-color: $bg-orange;
-  &:hover{
-    background-color: $bg-orange-hover;
-  }
-}
-
-.bg-green {
-  background-color: $bg-green;
-  &:hover{
-    background-color: $bg-green-hover;
-  }
-}
-
-.bg-pink {
-  background-color: $bg-pink;
-  &:hover{
-    background-color: $bg-pink-hover;
-  }
-}
-
-.bg-blue {
-  background-color: $bg-blue;
-  &:hover{
-    background-color: $bg-blue-hover;
-  }
-}
-
-.bg-purple {
-  background-color: $bg-purple;
-  &:hover{
-    background-color: $bg-purple-hover;
-  }
-}
-
-.bg-red {
-  background-color: $bg-red;
-  &:hover{
-    background-color: $bg-red-hover;
-  }
-}
-
-.bg-teal {
-  background-color: $bg-teal;
-  &:hover{
-    background-color: $bg-teal-hover;
-  }
 }
 
 .materia {
@@ -89,6 +17,13 @@
   width: 100%;
   height: 7vh;
   text-align: center;
+}
+
+.boton-color {
+  border: 1px solid;
+  color: white;
+  height: 2rem;
+  // width: 2rem;
 }
 
 @media (max-width:1461px){
